fixes and improvements
Fixed “Invalid field:WinAlpha” error in Windows Options state.
Added credits in CreditsState.
Added missing code that prevents player from running SLE in administrator mode (Windows).
Fixed a bug when compiling to Linux or MacOS (?).
Improved the way in which the player is told that he has an old version of SLE in the SlushiTitleState.

And minor changes.