NEW VERSION, NEW LUA DOCUMENTATION, MANY NEW THINGS!
This version is still based on the same stable version of SCE, based on commit 4eaf8c4.

- New way to use Windows GDI effects, now they can no longer lag the engine when used!
- New functions to tween the engine window border color, as if it was “FlxTween.color”.
- New FreeplayState and PauseSubState for SLE (although based on the original SCE ones)
- Minor changes in some engine details:
1. now the window will always be centered by Lime functions, not by C++ if it is with Windows.
2. the color of the border of the engine window changes in more states, it is something only aesthetic.

SLE now uses its fork of Flixel Addons of Glowsoony, without changes but, this way I assure that SLE can be compiled without problems if this library changes something.
Whatever bugs there may be with SCE, they are still there as it is still the same version as the previous version of SLE, I have not switched to the experimental version of SCE.